Move graph data structures to a new file, add new serialization version and adjust loading accordingly.
