+++ src/VecSim/algorithms/hnsw/hnsw.h
@@ -6,6 +6,7 @@
 
 #pragma once
 
+#include "graph_data.h"
 #include "visited_nodes_handler.h"
 #include "VecSim/spaces/spaces.h"
 #include "VecSim/memory/vecsim_malloc.h"
@@ -38,7 +39,6 @@
 
 using std::pair;
 
-typedef uint16_t linkListSize;
 typedef uint8_t elementFlags;
 
 template <typename DistType>
@@ -77,93 +77,6 @@ struct ElementMetaData {
         : label(label), flags(IN_PROCESS) {}
 };
 #pragma pack() // restore default packing
-
-// Helper method that swaps the last element in the ids list with the given one (equivalent to
-// removing the given element id from the list).
-bool removeIdFromList(vecsim_stl::vector<idType> &element_ids_list, idType element_id);
-
-struct LevelData {
-    // A list of ids that are pointing to the node where each edge is *unidirectional*
-    vecsim_stl::vector<idType> *incomingEdges;
-    // Total size of incoming links to the node (both uni and bi directinal).
-    linkListSize totalIncomingLikns;
-    linkListSize numLinks;
-    // Flexible array member - https://en.wikipedia.org/wiki/Flexible_array_member
-    // Using this trick, we can have the links list as part of the LevelData struct, and avoid
-    // the need to dereference a pointer to get to the links list.
-    // We have to calculate the size of the struct manually, as `sizeof(LevelData)` will not include
-    // this member. We do so in the constructor of the index, under the name `levelDataSize` (and
-    // `elementGraphDataSize`). Notice that this member must be the last member of the struct and
-    // all nesting structs.
-    idType links[];
-
-    explicit LevelData(std::shared_ptr<VecSimAllocator> allocator)
-        : incomingEdges(new (allocator) vecsim_stl::vector<idType>(allocator)),
-          totalIncomingLikns(0), numLinks(0) {}
-
-    linkListSize getNumLinks() const { return this->numLinks; }
-    idType getLinkAtPos(size_t pos) const {
-        assert(pos < numLinks);
-        return this->links[pos];
-    }
-    const vecsim_stl::vector<idType> &getIncomingEdges() const { return *incomingEdges; }
-    std::vector<idType> copyLinks() {
-        std::vector<idType> links_copy;
-        links_copy.assign(links, links + numLinks);
-        return links_copy;
-    }
-    // Sets the outgoing links of the current element.
-    // Assumes that the object has the capacity to hold all the links.
-    void setLinks(vecsim_stl::vector<idType> &links) {
-        numLinks = links.size();
-        memcpy(this->links, links.data(), numLinks * sizeof(idType));
-    }
-    template <typename DistType>
-    void setLinks(candidatesList<DistType> &links) {
-        numLinks = 0;
-        for (auto &link : links) {
-            this->links[numLinks++] = link.second;
-        }
-    }
-    void popLink() { this->numLinks--; }
-    void setNumLinks(linkListSize num) { this->numLinks = num; }
-    void setLinkAtPos(size_t pos, idType node_id) { this->links[pos] = node_id; }
-    void appendLink(idType node_id) { this->links[this->numLinks++] = node_id; }
-    void newIncomingUnidirectionalEdge(idType node_id) { this->incomingEdges->push_back(node_id); }
-    bool removeIncomingUnidirectionalEdgeIfExists(idType node_id) {
-        return removeIdFromList(*this->incomingEdges, node_id);
-    }
-    void increaseTotalIncomingEdgesNum() { this->totalIncomingLikns++; }
-    void decreaseTotalIncomingEdgesNum() { this->totalIncomingLikns--; }
-    void swapNodeIdInIncomingEdges(idType id_before, idType id_after) {
-        auto it = std::find(this->incomingEdges->begin(), this->incomingEdges->end(), id_before);
-        // This should always succeed
-        assert(it != this->incomingEdges->end());
-        *it = id_after;
-    }
-};
-
-struct ElementGraphData {
-    size_t toplevel;
-    std::mutex neighborsGuard;
-    LevelData *others;
-    LevelData level0;
-
-    ElementGraphData(size_t maxLevel, size_t high_level_size,
-                     std::shared_ptr<VecSimAllocator> allocator)
-        : toplevel(maxLevel), others(nullptr), level0(allocator) {
-        if (toplevel > 0) {
-            others = (LevelData *)allocator->callocate(high_level_size * toplevel);
-            if (others == nullptr) {
-                throw std::runtime_error("VecSim index low memory error");
-            }
-            for (size_t i = 0; i < maxLevel; i++) {
-                new ((char *)others + i * high_level_size) LevelData(allocator);
-            }
-        }
-    }
-    ~ElementGraphData() = delete; // Should be destroyed using `destroyGraphData`
-};
 
 //////////////////////////////////// HNSW index implementation ////////////////////////////////////
 
@@ -1693,20 +1606,10 @@ void HNSWIndex<DataType, DistType>::insertElementToGraph(idType element_id,
         max_common_level = global_max_level;
     }
 
-    size_t levels_without_neighbors = 0;
     for (auto level = static_cast<int>(max_common_level); level >= 0; level--) {
         candidatesMaxHeap<DistType> top_candidates =
-            searchLayer<true>(curr_element, vector_data, level, efConstruction);
-        // If the entry point was marked deleted between iterations, we may recieve an empty
-        // candidates set.
-        if (!top_candidates.empty()) {
-            curr_element = mutuallyConnectNewElement(element_id, top_candidates, level);
-        } else {
-            levels_without_neighbors++;
-        }
-    }
-    if (levels_without_neighbors == max_common_level + 1) {
-        std::cout << "Empty candidates set" << std::endl;
+            searchLayer<false>(curr_element, vector_data, level, efConstruction);
+        curr_element = mutuallyConnectNewElement(element_id, top_candidates, level);
     }
 }
