+++ src/VecSim/algorithms/hnsw/hnsw_serializer.h
@@ -100,11 +100,12 @@ HNSWIndexMetaData HNSWIndex<DataType, DistType>::checkIntegrity() const {
         return res;
     }
 
-    // Validate that in-degree out each node is coherent with
+    // Validate that each node's in-degree is coherent with the in-degree observed by the
+    // outgoing edges.
     for (size_t i = 0; i < this->curElementCount; i++) {
         for (size_t l = 0; l <= getGraphDataByInternalId(i)->toplevel; l++) {
             LevelData &cur = this->getLevelData(i, l);
-            if (cur.totalIncomingLikns != inbound_connections_num[i][l]) {
+            if (cur.totalIncomingLinks != inbound_connections_num[i][l]) {
                 return res;
             }
             if (inbound_connections_num[i][l] > max_in_degree) {
@@ -152,7 +153,7 @@ void HNSWIndex<DataType, DistType>::restoreIndexFields(std::ifstream &input) {
 }
 
 template <typename DataType, typename DistType>
-void HNSWIndex<DataType, DistType>::restoreGraph(std::ifstream &input) {
+void HNSWIndex<DataType, DistType>::restoreGraph(std::ifstream &input, EncodingVersion version) {
     // Restore id to metadata vector
     labelType label = 0;
     elementFlags flags = 0;
@@ -213,17 +214,50 @@ void HNSWIndex<DataType, DistType>::restoreGraph(std::ifstream &input) {
 
             // Restore the current element's graph data
             for (size_t k = 0; k <= toplevel; k++) {
-                restoreLevel(input, getLevelData(cur_egt, k));
+                restoreLevel(input, getLevelData(cur_egt, k), version);
             }
+        }
+    }
+    if (version < EncodingVersion_V4) {
+        this->computeIndegreeForAll();
+    }
+}
+
+template <typename DataType, typename DistType>
+void HNSWIndex<DataType, DistType>::computeIndegreeForAll() {
+    size_t max_level_in_graph = 0; // including marked deleted elements
+    for (size_t i = 0; i < this->curElementCount; i++) {
+        if (getGraphDataByInternalId(i)->toplevel > max_level_in_graph) {
+            max_level_in_graph = getGraphDataByInternalId(i)->toplevel;
+        }
+    }
+    std::vector<std::vector<int>> inbound_connections_num(
+        this->curElementCount, std::vector<int>(max_level_in_graph + 1, 0));
+    for (size_t i = 0; i < this->curElementCount; i++) {
+        for (size_t l = 0; l <= getGraphDataByInternalId(i)->toplevel; l++) {
+            LevelData &cur = this->getLevelData(i, l);
+            std::set<idType> s;
+            for (unsigned int j = 0; j < cur.numLinks; j++) {
+                inbound_connections_num[cur.links[j]][l]++;
+            }
+        }
+    }
+    // Populate the total incoming links for each node.
+    for (size_t i = 0; i < this->curElementCount; i++) {
+        for (size_t l = 0; l <= getGraphDataByInternalId(i)->toplevel; l++) {
+            LevelData &cur = this->getLevelData(i, l);
+            cur.totalIncomingLinks = inbound_connections_num[i][l];
         }
     }
 }
 
 template <typename DataType, typename DistType>
-void HNSWIndex<DataType, DistType>::restoreLevel(std::ifstream &input, LevelData &data) {
+void HNSWIndex<DataType, DistType>::restoreLevel(std::ifstream &input, LevelData &data,
+                                                 EncodingVersion version) {
     // Restore the links of the current element
-    readBinaryPOD(input,
-                  data.totalIncomingLikns); // todo: this does not exist in serialzed benchmarks
+    if (version >= EncodingVersion_V4) {
+        readBinaryPOD(input, data.totalIncomingLinks);
+    }
     readBinaryPOD(input, data.numLinks);
     for (size_t i = 0; i < data.numLinks; i++) {
         readBinaryPOD(input, data.links[i]);
@@ -317,7 +351,7 @@ void HNSWIndex<DataType, DistType>::saveGraph(std::ofstream &output) const {
 template <typename DataType, typename DistType>
 void HNSWIndex<DataType, DistType>::saveLevel(std::ofstream &output, LevelData &data) const {
     // Save the links of the current element
-    writeBinaryPOD(output, data.totalIncomingLikns);
+    writeBinaryPOD(output, data.totalIncomingLinks);
     writeBinaryPOD(output, data.numLinks);
     for (size_t i = 0; i < data.numLinks; i++) {
         writeBinaryPOD(output, data.links[i]);
